Setup closure builder and fixed closure compiler
Setup the closure builder and replaced it as the concat task. This
will combine all the required good files with the src files into one
aronnax js file in the dist folder. This file may have to be used for
the tests, otherwise not all dependencies from goog will be there. It
could possibly be setup by providing multiple sources in the jasmine
conf, I will attempt to do so. It would be best to test off individual
source files rather then the built source as it will be easier to fix
mistakes

Other
- Put in a good export symbol but not sure if it's required