Added benchmark js suite into project but not build
Adding the benchmark suite to the build would require a new grunt task
to be completely built. This would have to create a default html page
with all the correct scripts included. It would then need a reporter
which could take the benchmark.stats object, pull the relevant stats
for the test and print them to a report file. The hard part is that
this would have to run for all benchmark tests that need to be run.

Other changes:
- Updated package.json with all new depends
- Fixed formatting issues in grunt file
- Made jasmine tests to not recursively search sub directories in
tests folder as there is now a benchmark folder