Added full docs to accessor mixin and test attrs method
Added full docs to the accessor mixin. This is set as a mixin in the
docs. The LinkedList classes have to be modified so they are not
considered namespaces. The main class now holds the reference to the
aronnax namespace. The closure compiler errors about a bad global this
variable was fixed by setting the @this param in the doc to the
accessor object. This has to be done on the Linked Lists. The test for
the attrs method was created, it uses a test constructor class so it
correctly tests the correct object attributes.

Other changes:
- Ran full build
- Added aronnax namespace to main