Added logger class
Logger class will take an string or object and maps directly to
browser console functions, log, warn, error. The object could be
a standard error object, with message, url, line.

Currently the code just prints the message through console, but will
be refactored later to send the message to the server.

Other changes:
- Ran full dist
- Put in exception in Lint to create functions in a loop