Added dependency injection for logger
The logger should not depend on the implementation of what logger
is to be used, in case a different logger should be used in
different environments, or a third-party logger should be used.
This sets the dependency for the logger in the Gruntfile so it can
be compiled in different ways.