Refactor logger with dependencies Fixes #37 and fixes #38
Moves the logger logwriter dependency to the test files so it
doesn't use the acutal window.console. This is done by using the
require.config in the test file and setting the logWriter to a
mock object.

Also moves the logwriter dep to the Log class rather then the logger
class because that makes more sense.