Modified Base create so it allows functions to be passed in
I feel this is somewhat controversial, due to mapping to changing
the mapping to the Object.create function in a manner that does
not conform to the Javascript specifications. I also feel that
its much easier to write code with both functions and properties
passed into the Base.create function like this.