Got the Pooled prototype up and running Fixes #34
The pooled prototype is working by returning the correct object
with the make function. The release function stil has to be written.

The actual calling class of the make function is passed to Pool, and
then this prototype is stored as a property on the Pool. It's used
when creating the objects for the Pool.

One issue I had with acessors, I wasn't able to have get and set
function working within an object, while not using a private var.