Added integration tests
These tests will test out various components are acting together.
Right now they are running in jasmine but may need to be expanded
later to incorperate other systems.

The organization of the tests is off right now. They should be in
separate files but under the same namespace if possible.

Other changes:
- util: put in a try block to ensure only writable properties were
deleted.
- store: ensure object is returned in the remove function.