Fixed documentation
Changed the documentation to plain jsdoc rather then the AMD js doc
modules which was not working correctly.

Changed all files so they use the same pattern as exporting a module
and have a main class. For prototypes, the it lends the prototype of the
main class and doesn't include a class definition of its own. Also setup
extending. Have to fix Base extend.

TODO
- Change base so it has a proto to inherit from
- Put all things in Aronnax class
- If possible find a way to have the modules show the actually classes
  they are exporting.

Other changes:
- Ran build
- Updated README